File Handling
This practice deals with the:
Reading the file.
writing into the file.
Appending into the file.
Some file operations like rename, getting the size, creating the copy of the file, deleting the file.
File operations by using the "with" command.
Read and dumping the details into json file.
Reading and writing the csv file.
Reading and writing of the binary file.
Buffered read and write operations.